requirements.txt
PyPDF2
pdfminer.six==20181108
attrs==17.4.0
certifi==2018.1.18
cryptography==2.1.4
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
MarkupSafe==1.0
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
service-identity==16.0.0
six==1.11.0
urllib3==1.22
